style: responsive layout
